+++ KalikoCMS.Admin/Admin/Content/EditPage.aspx.cs
@@ -17,11 +17,11 @@
  */
 #endregion
 
-using KalikoCMS.Data;
-
 namespace KalikoCMS.Admin.Content {
     using System;
     using System.Collections.Generic;
+    using System.Web.UI.WebControls;
+    using Data;
     using Core;
     using Extensions;
     using KalikoCMS.PropertyType;
@@ -77,6 +77,8 @@ namespace KalikoCMS.Admin.Content {
         }
 
         private void LoadControls() {
+            LoadChildSortOrderLists();
+
             if (_pageId != Guid.Empty) {
                 LoadFormForExistingPage();
             }
@@ -86,6 +88,17 @@ namespace KalikoCMS.Admin.Content {
             else {
                 LoadFormForRootPage();
             }
+        }
+
+        private void LoadChildSortOrderLists() {
+            ChildSortOrder.Items.Add(new ListItem("Created date", ((int)Core.Collections.SortOrder.CreatedDate).ToString()));
+            ChildSortOrder.Items.Add(new ListItem("Page name", ((int)Core.Collections.SortOrder.PageName).ToString()));
+            ChildSortOrder.Items.Add(new ListItem("Sort index", ((int)Core.Collections.SortOrder.SortIndex).ToString()));
+            ChildSortOrder.Items.Add(new ListItem("Start pubish date", ((int)Core.Collections.SortOrder.StartPublishDate).ToString()));
+            ChildSortOrder.Items.Add(new ListItem("Updated date", ((int)Core.Collections.SortOrder.UpdateDate).ToString()));
+
+            ChildSortDirection.Items.Add(new ListItem("Ascending", ((int)Core.Collections.SortDirection.Ascending).ToString()));
+            ChildSortDirection.Items.Add(new ListItem("Descending", ((int)Core.Collections.SortDirection.Descending).ToString()));
         }
 
         private void LoadFormForRootPage() {
@@ -103,10 +116,13 @@ namespace KalikoCMS.Admin.Content {
             SetStandardFieldLabels();
 
             var propertyDefinitions = PageType.GetPropertyDefinitions(_pageTypeId);
-
             foreach (var propertyDefinition in propertyDefinitions) {
                 AddControl(propertyDefinition.Name, null, propertyDefinition.PropertyTypeId, propertyDefinition.Header, propertyDefinition.Parameters);
             }
+
+            var pageType = PageType.GetPageType(_pageTypeId);
+            ChildSortDirection.SelectedValue = ((int)pageType.DefaultChildSortDirection).ToString();
+            ChildSortOrder.SelectedValue = ((int)pageType.DefaultChildSortOrder).ToString();
         }
 
         private void SetStandardFieldLabels() {
@@ -115,7 +131,6 @@ namespace KalikoCMS.Admin.Content {
             StopPublishDate.PropertyLabel = "Stop publish";
             VisibleInMenu.PropertyLabel = "Visible in menus";
             VisibleInSitemap.PropertyLabel = "Visible in sitemaps";
-            SortOrder.PropertyLabel = "Sort order";
         }
 
         private void LoadFormForExistingPage() {
@@ -153,8 +168,9 @@ namespace KalikoCMS.Admin.Content {
             // Advanced fields
             VisibleInSitemap.PropertyValue = new BooleanProperty(cmsPage.VisibleInSiteMap);
             PageUrlSegment.Text = cmsPage.UrlSegment;
-            SortOrder.PropertyValue = new NumericProperty(cmsPage.SortOrder);
             OldPageUrlSegment.Value = cmsPage.UrlSegment;
+            ChildSortDirection.SelectedValue = ((int)cmsPage.ChildSortDirection).ToString();
+            ChildSortOrder.SelectedValue = ((int)cmsPage.ChildSortOrder).ToString();
 
             PageId.Text = cmsPage.PageId.ToString();
 
@@ -168,13 +184,11 @@ namespace KalikoCMS.Admin.Content {
             }
         }
 
-
         private void SaveButtonEventHandler(object sender, EventArgs e) {
             if(IsDataValid) {
                 var page = SaveData();
 
                 if (_pageTypeId > 0) {
-                    page.Publish(true);
                     Feedback.Text = string.Format("<script>parent.$('.notifications.top-right').notify({{ type: 'info', message: \"<i class=\\\"icon-flag\\\"></i> Page <b>{0}</b> has been created!!\", fadeOut: {{ enabled: true, delay: 5000 }}}}).show();parent.refreshTreeNode('{1}','{2}');document.location = '{3}?id={2}';</script>", _pageName, _parentId, _pageId, Request.Path);
                     Feedback.Visible = true;
                 }
@@ -190,6 +204,7 @@ namespace KalikoCMS.Admin.Content {
         private void PublishButtonEventHandler(object sender, EventArgs e) {
             if(IsDataValid) {
                 var page = SaveData();
+
                 page.Publish();
 
                 ScriptArea.Text = "top.refreshNode('" + CurrentPageId + "');";
@@ -199,7 +214,8 @@ namespace KalikoCMS.Admin.Content {
                     Feedback.Visible = true;
                 }
                 else {
-                    ShowMessage(Feedback, String.Format("Page <b>{0}</b> has been published!", _pageName));
+                    Feedback.Text = string.Format("<script>parent.$('.notifications.top-right').notify({{ type: 'info', message: \"<i class=\\\"icon-flag\\\"></i> Page <b>{0}</b> has been published!!\", fadeOut: {{ enabled: true, delay: 5000 }}}}).show();</script>", _pageName);
+                    Feedback.Visible = true;
                 }
             }
             else {
@@ -218,9 +234,6 @@ namespace KalikoCMS.Admin.Content {
                     valid = false;
                 }
                 else if (!StopPublishDate.Validate()) {
-                    valid = false;
-                }
-                else if (!SortOrder.Validate()) {
                     valid = false;
                 }
 
@@ -263,7 +276,8 @@ namespace KalikoCMS.Admin.Content {
             editablePage.SetStopPublish(((DateTimeProperty)StopPublishDate.PropertyValue).Value);
             editablePage.SetVisibleInMenu(((BooleanProperty) VisibleInMenu.PropertyValue).Value);
             editablePage.SetVisibleInSiteMap(((BooleanProperty)VisibleInSitemap.PropertyValue).Value);
-            editablePage.SetSortOrder(((NumericProperty)SortOrder.PropertyValue).Value);
+            editablePage.SetChildSortDirection(int.Parse(ChildSortDirection.SelectedValue));
+            editablePage.SetChildSortOrder(int.Parse(ChildSortOrder.SelectedValue)); 
 
             HandlePageUrlSegment(editablePage);