+++ KalikoCMS.Admin/Admin/Content/PageTree/JQueryTreeContent.ashx.cs
@@ -21,8 +21,9 @@ namespace KalikoCMS.Admin.Content.PageTree {
     using System;
     using System.Text;
     using System.Web;
-    using Core;
-    using Data;
+    using KalikoCMS.Core;
+    using KalikoCMS.Data;
+    using KalikoCMS.Caching;
 
     public class JQueryTreeContent : IHttpHandler {
         private void GetChildren(HttpContext context) {
@@ -36,7 +37,6 @@ namespace KalikoCMS.Admin.Content.PageTree {
 
             var pageId = new Guid(id);
             var separator = string.Empty;
-
 
             var children = PageFactory.GetChildrenForPage(pageId, PublishState.All);
             var stringBuilder = new StringBuilder();
@@ -59,11 +59,31 @@ namespace KalikoCMS.Admin.Content.PageTree {
         private void MoveNode(HttpContext context) {
             var pageId = new Guid(context.Request.Form["id"]);
             var targetId = new Guid(context.Request.Form["ref"]);
+            var position = int.Parse(context.Request.Form["position"] ?? "0");
+            var oldParentId = new Guid(context.Request.Form["old"]);
 
-            PageFactory.MovePage(pageId, targetId);
+            CacheManager.RemoveRelated(targetId);
+            CacheManager.RemoveRelated(oldParentId);
 
+            if (targetId == oldParentId) {
+                var success = PageFactory.ReorderChildren(pageId, targetId, position);
+                if (success) {
+                    WriteResponse(context, true, string.Empty);
+                }
+                else {
+                    WriteResponse(context, false, "Parent is not set to allow manual resorting!");
+                }
+            }
+            else {
+                PageFactory.MovePage(pageId, targetId, position);
+                WriteResponse(context, true, string.Empty);
+            }
+        }
+
+        private static void WriteResponse(HttpContext context, bool status, string message) {
+            var statusString = status ? "true" : "false";
             context.Response.ContentType = "application/json";
-            context.Response.Write("{ \"success\": true }");
+            context.Response.Write(string.Format("{{ \"success\": {0}, \"message\": \"{1}\" }}", statusString, message));
         }
 
         private void RemoveNode(HttpContext context) {