+++ KalikoCMS.Engine/Data/AutoMapperConfiguration.cs
@@ -41,6 +41,9 @@ namespace KalikoCMS.Data {
                 .ReverseMap();
             
             Mapper.CreateMap<PageTypeEntity, PageType>()
+                .ForMember(m => m.AllowedTypes, o => o.Ignore())
+                .ForMember(m => m.Type, o => o.Ignore())
+                .ForMember(m => m.PreviewImage, o => o.Ignore())
                 .ReverseMap();
 
             Mapper.CreateMap<PropertyItem, PropertyItem>();